Move date related functions into date.go
